domu/requirements.txt
Flask==1.1.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Werkzeug==1.0.1
Jinja2==2.11.3
MarkupSafe==2.0.1  # Last version with soft_unicode
psycopg2-binary==2.9.3
python-dotenv==0.19.0